requirements.txt
numpy
requests
python-dotenv
tqdm
streamlit
huggingface_hub
FlagEmbedding
confz
opencv-python
pyyaml
pillow
openai
streamlit_cropper